+++ CHANGELOG.md
@@ -4,15 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - [unreleased]
 
+## [1.1.0] - 2018-06-06
+
 ### Added
 - Wikipedia's 1000 most viewed articles for any given month or day now accessible via the `wikipedia-trends` task!
 - With bug fix for wikipedia-views (see below), there is now better support for daily page view statistics from wikipedia. This includes day-by-day and even hour-by-hour support thanks to wikimedia's API. This changes the call syntax for wikipedia-views (changes are reflected in `massmine -h task-options`)
+- Unit tests! MassMine can now run tests for each module. Tests are included in the tests directory. Example: ./massmine.scm --test ./tests/run.scm
 
 ### Fixed
 - wikipedia-views task updated to fix broken functionality. Now fetches data from the wikimedia API. 
-
-### Added
-- Unit tests! MassMine can now run tests for each module. Tests are included in the tests directory. Example: ./massmine.scm --test ./tests/run.scm
 
 ## [1.0.2] - 2016-09-07